# imaware_devops_challenge/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Imaware Tech Challenge API
components:
  schemas:
    Observation:
      description: This represents an observation datum about a Patient.
      properties:
        observationId:
          type: string
          format: uuid
          description: The UUID of the Observation.
        observationType:
          type: string
          enum:
            - stringObservation
            - booleanObservation
            - integerObservation
          description: The type of Observation data.
        observationDateTime:
          type: string
          format: date
          description: The datetime timestamp the Observation was taken (ISO 8601 format).
        valueString:
          type: string
          description: A string value for an Observation.
        valueBoolean:
          type: boolean
          description: A boolean value for an Observation.
        valueInteger:
          type: number
          description: An integer value for an Observation.
        patientId:
          type: string
          format: uuid
          description: The UUID of the associated Patient.
      required:
        - observationId
        - observationType
        - observationDateTime
        - patientId
paths:
  '/patientData/{id}':
    get:
      summary: Search for patient data
      description: Returns all patient data matching the search criteria.
      operationId: getPatientObservations
      parameters:
        - name: id
          in: path
          description: The Patient's UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: observationType
          in: query
          description: The Observation type
          schema:
            type: string
            enum:
              - stringObservation
              - booleanObservation
              - integerObservation
        - name: startDate
          in: query
          description: The minimum date for the Observation's timestamp (ISO 8601 format).
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: The maximum date for the Observation's timestamp (ISO 8601 format).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Observation'
        '204':
          description: No results found for given search parameters
        '404':
          description: No patient found for provided id
        '500':
          description: Internal server error